Trigger, Gate, Bridge & Over A rows and context menu (randomize all)
___trigger, gate, bridge & over a
in: features
"There is a trigger output jack for each row. A trigger is fired the instant a step/row is activated. If repeats are active, the trigger will fire a ratchet. 
It will operate in one of four modes, depending on the setting of the lamp in the mode column for that row.

The gate row will send a gate for as long as the row is active. 

'Bridge' pairs an input and output jack. Input can be any monophonic signal (in- or external i.e.: audio, cv, clk/trg/gate). The input signal will be sent to the output jack only while the corresponding channel is active. A possible scenario would be to bind a clock signal/trigger sequence to (a) specific step(s): advance another sequence, activate an arpeggio, trigger a drumfill. You could also send audio into effects. 

'Over A' takes (like 'Bridge') an monophonic voltage for input. 'Over A's input overwrites column A's knob and chain in values for its respective row. 
While the row is active, 'Over A's input is sent directly to column As output. Add random voltages to some steps of column A for example. You could use the whole column as a switch: take the trigger/gate outputs of each row to activate various envelopes and send these to 'Over A'. "

//////////////////////////////////////

___context menu
in: the lower area
"The sections affected by 'randomize all' can be specified in the context menu (right-click on Squonk). CV Line A, B, C, D, E, Repetitions, Direction, x5 multiplier and MODE can be turned on/off. Per default all sections are turned on. "